useActionExerciseSession 뮤테이션 훅과 api/user/exerciseSession/[sessionId]/[id]/route.ts 의 POST 코드 개선 개선 내용 req로 받는 데이터를     sessionId: string,     exerciseId: string,     action: "start" | "done", 이 형식으로 교체함 내부 로직은 action에 따라 결정됨 이제부턴 클라이언트에서 세트를 생성하는 것이 아닌 서버에서 세트를 직접 생성함 또한 완료 시간 도 서버에서 측정하도록 변경
