+++ app/dashboard/page.tsx
@@ -1,25 +1,11 @@
 import ShowChart from "@/components/UserCpmponents/showChart";
 import ShowPlans from "@/components/UserCpmponents/showPlans";
 import ShowWeek from "@/components/UserCpmponents/showWeek";
+import { fetchWithCookie } from "@/utils/fetchUrl";
 import { HydrationBoundary, dehydrate, QueryClient } from "@tanstack/react-query";
 import { cookies } from "next/headers";
 
-const fetchWithCookie = async (url: string, cookieName: string, cookieValue: string | undefined) => {
-  const response = await fetch(url, {
-    method: "GET",
-    headers: {
-      "Content-Type": "application/json",
-      "Cookie": cookieValue ? `${cookieName}=${cookieValue}` : "",
-    },
-    credentials: "include",
-  });
 
-  if (!response.ok) {
-    throw new Error(`HTTP error! status: ${response.status}`);
-  }
-
-  return response.json();
-};
 
 const fetchData = async () => {
   const cookieHeader = await cookies();
