+++ providers/QueryProvider.tsx
@@ -10,8 +10,7 @@ export function QueryProviders({ children }: { children: React.ReactNode }) {
     const [queryClient] = useState(() => new QueryClient({
         defaultOptions: {
             queries: {
-
-
+                staleTime: 1000 * 60 * 5, // 5분
             },
         },
     })); // useState로 생성
